Issue #1046 move render service item function to its own file for re-use
