+++ src/components/bookmarks/service_bookmarks_component.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
-import * as R from 'ramda';
 import { FlatList } from 'react-native';
 import { HumanServiceData } from '../../validation/services/types';
-import { ServiceItemInfo } from '../services/service_list_component';
 import { Trans } from '@lingui/react';
-import { RouterProps, goToRouteWithParameter, Routes } from '../../application/routing';
-import { ServiceListItemComponent } from '../services/service_list_item_component';
+import { RouterProps } from '../../application/routing';
 import { colors } from '../../application/styles';
 import { View } from 'native-base';
 import { EmptyBookmarksComponent } from './empty_bookmarks_component';
 import { BookmarkServiceAction, UnbookmarkServiceAction } from '../../stores/services/actions';
+import { renderServiceItems } from '../services/render_service_items';
 
 export interface ServiceBookmarksProps {
     readonly bookmarkedServices: ReadonlyArray<HumanServiceData>;
@@ -36,20 +34,3 @@ export const ServiceBookmarksComponent = (props: Props): JSX.Element => (
         ListHeaderComponent={<View />}
     />
 );
-
-export const renderServiceItems = R.curry((props: Props, itemInfo: ServiceItemInfo): JSX.Element => {
-    const service = itemInfo.item;
-    const onBookmark = (): BookmarkServiceAction => props.bookmarkService(service);
-    const onUnbookmark = (): UnbookmarkServiceAction => props.unbookmarkService(service);
-    return (
-        <ServiceListItemComponent
-            history={props.history}
-            service={service}
-            onPress={goToRouteWithParameter(Routes.ServiceDetail, service.id, props.history)}
-            currentPath={props.location.pathname}
-            isBookmarked={service.bookmarked}
-            onBookmark={onBookmark}
-            onUnbookmark={onUnbookmark}
-        />
-    );
-});
