+++ src/components/services/render_service_items.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import * as R from 'ramda';
+import { goToRouteWithParameter, Routes } from '../../application/routing';
+import { BookmarkServiceAction, UnbookmarkServiceAction } from '../../stores/services/actions';
+import { ServiceListItemComponent } from './service_list_item_component';
+import { ListRenderItemInfo } from 'react-native';
+import { HumanServiceData } from '../../validation/services/types';
+import { History } from 'history';
+
+export interface ServiceItemsProps {
+    readonly history: History;
+    readonly bookmarkService: (service: HumanServiceData) => BookmarkServiceAction;
+    readonly unbookmarkService: (service: HumanServiceData) => UnbookmarkServiceAction;
+}
+
+export type ServiceItemInfo = ListRenderItemInfo<HumanServiceData>;
+
+export const renderServiceItems = R.curry((props: ServiceItemsProps, itemInfo: ServiceItemInfo): JSX.Element => {
+    const service = itemInfo.item;
+    const onBookmark = (): BookmarkServiceAction => props.bookmarkService(service);
+    const onUnbookmark = (): UnbookmarkServiceAction => props.unbookmarkService(service);
+    return (
+        <ServiceListItemComponent
+            history={props.history}
+            service={service}
+            onPress={goToRouteWithParameter(Routes.ServiceDetail, service.id, props.history)}
+            isBookmarked={service.bookmarked}
+            onBookmark={onBookmark}
+            onUnbookmark={onUnbookmark}
+        />
+    );
+});
